rename options for select to constants
